requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.4
discord.py==0.16.8
idna==2.5
multidict==3.1.0
urllib3==1.21.1
websockets==3.3
